requirements.txt
gradio==5.18.0
matplotlib==3.10.0
numpy==2.2.3
pandas==2.2.3
Pillow==11.1.0
tifffile==2025.1.10
torch==2.5.0
torchvision==0.20.0
tqdm==4.67.1
scikit-image
scikit-learn>=1.3
pynrrd
opencv-python==4.11.0.86
